initial commit
split the microcontroller entrypoint out of the gotogen project so that project itself does not technically need the tinygo libraries. that effort may not be complete yet, but I want to get to that point because GoLand isn't happy with tinygo on Windows for me, and fixing that would make my development life much easier with the simulator.